requirements.txt
ttkbootstrap
pillow
python-dotenv
mysql-connector-python
opencv-python
face_recognition